+++ ta4j-core/src/main/java/org/ta4j/core/analysis/cost/FixedTransactionCostModel.java
@@ -29,8 +29,6 @@ import org.ta4j.core.num.Num;
 
 public class FixedTransactionCostModel implements CostModel {
 
-    private static final long serialVersionUID = 3486293523619720786L;
-
     /**
      * Cost per {@link Trade trade}.
      */