+++ ta4j-core/src/main/java/org/ta4j/core/analysis/cost/LinearBorrowingCostModel.java
@@ -29,7 +29,6 @@ import org.ta4j.core.num.Num;
 
 public class LinearBorrowingCostModel implements CostModel {
 
-    private static final long serialVersionUID = -2839623394737567618L;
     /**
      * Slope of the linear model - fee per period
      */