+++ ta4j-core/src/main/java/org/ta4j/core/analysis/cost/ZeroCostModel.java
@@ -25,15 +25,12 @@ package org.ta4j.core.analysis.cost;
 
 public class ZeroCostModel extends FixedTransactionCostModel {
 
-    private static final long serialVersionUID = -331246208177262596L;
-
     private static final double ZERO_FEE_PER_TRADE = 0.0;
 
     /**
      * Constructor for a trading cost-free model.
      *
      * @see FixedTransactionCostModel fixedTransactionCostModel
-     *
      */
     public ZeroCostModel() {
         super(ZERO_FEE_PER_TRADE);